# perf(sql): use JIT compiled filter in parallel GROUP BY