--- a/core/src/main/java/io/questdb/cairo/map/Map.java
+++ b/core/src/main/java/io/questdb/cairo/map/Map.java
@@ -37,6 +37,11 @@ public interface Map extends Mutable, Closeable, Reopenable {
 
     MapRecordCursor getCursor();
 
+    @TestOnly
+    default long getHeapSize() {
+        return -1;
+    }
+
     /**
      * Returns full (physical) key capacity of the map ignoring the load factor.
      */
@@ -45,7 +50,21 @@ public interface Map extends Mutable, Closeable, Reopenable {
 
     MapRecord getRecord();
 
+    /**
+     * Returns used heap size in bytes for maps that store keys and values separately like {@link OrderedMap}.
+     * Returns -1 if the map doesn't use heap.
+     */
+    default long getUsedHeapSize() {
+        return -1;
+    }
+
     void merge(Map srcMap, MapValueMergeFunction mergeFunc);
+
+    /**
+     * Reopens previously closed map with given key capacity and page size.
+     * Page size value is ignored if the map does not use heap to store keys and values.
+     */
+    void reopen(int keyCapacity, int pageSize);
 
     void restoreInitialCapacity();