--- a/core/src/main/java/io/questdb/cairo/map/Unordered16Map.java
+++ b/core/src/main/java/io/questdb/cairo/map/Unordered16Map.java
@@ -69,10 +69,9 @@ public class Unordered16Map implements Map, Reopenable {
 
     static final long KEY_SIZE = 2 * Long.BYTES;
     private static final long MAX_SAFE_INT_POW_2 = 1L << 31;
-    private static final int MIN_INITIAL_CAPACITY = 16;
+    private static final int MIN_KEY_CAPACITY = 16;
     private final Unordered16MapCursor cursor;
     private final long entrySize;
-    private final int initialKeyCapacity;
     private final Key key;
     private final double loadFactor;
     private final int maxResizes;
@@ -83,6 +82,7 @@ public class Unordered16Map implements Map, Reopenable {
     private final Unordered16MapValue value3;
     private int free;
     private boolean hasZero;
+    private int initialKeyCapacity;
     private int keyCapacity;
     private long keyMemStart; // Key look-up memory start pointer.
     private int mask;
@@ -115,7 +115,7 @@ public class Unordered16Map implements Map, Reopenable {
         this.memoryTag = memoryTag;
         this.loadFactor = loadFactor;
         this.keyCapacity = (int) (keyCapacity / loadFactor);
-        this.keyCapacity = this.initialKeyCapacity = Math.max(Numbers.ceilPow2(this.keyCapacity), MIN_INITIAL_CAPACITY);
+        this.keyCapacity = this.initialKeyCapacity = Math.max(Numbers.ceilPow2(this.keyCapacity), MIN_KEY_CAPACITY);
         this.maxResizes = maxResizes;
         mask = this.keyCapacity - 1;
         free = (int) (this.keyCapacity * loadFactor);
@@ -267,6 +267,15 @@ public class Unordered16Map implements Map, Reopenable {
             if (--free == 0) {
                 rehash();
             }
+        }
+    }
+
+    @Override
+    public void reopen(int keyCapacity, int pageSize) {
+        if (memStart == 0) {
+            keyCapacity = (int) (keyCapacity / loadFactor);
+            initialKeyCapacity = Math.max(Numbers.ceilPow2(keyCapacity), MIN_KEY_CAPACITY);
+            restoreInitialCapacity();
         }
     }