@ideoma 2024-01-17
#### [PR Coverage check]

:heart_eyes: **pass** : 239 / 246 (97.15%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LimitedSizeSortedLightRecordCursorFactory.java|1|2|50.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncJitFilteredRecordCursorFactory.java|44|49|89.80%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByNotKeyedRecordCursorFactory.java|57|58|98.28%|
|:large_blue_circle:|io/questdb/cairo/sql/PageAddressCacheRecord.java|23|23|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/StealableFilterRecordCursorFactory.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/bool/OrFunctionFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByAtom.java|18|18|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByRecordCursorFactory.java|63|63|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|10|10|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByNotKeyedAtom.java|18|18|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/bool/AndFunctionFactory.java|1|1|100.00%|