@ideoma 2024-01-18
#### [PR Coverage check]

:heart_eyes: **pass** : 268 / 281 (95.37%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinCharGroupByFunction.java|2|7|28.57%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxCharGroupByFunction.java|1|2|50.00%|
|:large_blue_circle:|io/questdb/griffin/engine/orderby/LimitedSizeSortedLightRecordCursorFactory.java|1|2|50.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncJitFilteredRecordCursorFactory.java|44|49|89.80%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByNotKeyedRecordCursorFactory.java|57|58|98.28%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxDoubleGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/VwapDoubleGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/SumLongGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/sql/PageAddressCacheRecord.java|23|23|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/StealableFilterRecordCursorFactory.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinTimestampGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxFloatGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/bool/OrFunctionFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinFloatGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinIPv4GroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinDoubleGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByAtom.java|18|18|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxLongGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/AvgDoubleGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinDateGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxIPv4GroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxTimestampGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/AbstractCountGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxDateGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByRecordCursorFactory.java|63|63|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctIPv4GroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxStrGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinLongGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctLongGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinIntGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/SumDoubleGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinStrGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|10|10|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/SumIntGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/SumFloatGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByNotKeyedAtom.java|18|18|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxIntGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/CountDistinctIntGroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/bool/AndFunctionFactory.java|1|1|100.00%|