@puzpuzpuz 2024-01-18
> it is not a huge deal, but we seem to have regressed here

The difference is small, I believe it's due to jitter. Would be great if you could run the query locally before and after - this way we'll have more results.